pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tinbox"
version = "0.1.0"
description = "A CLI translation tool using LLMs for document translation"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = ["translation", "llm", "cli", "pdf", "docx"]
authors = [{ name = "strickvl", email = "your.email@example.com" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pydantic>=2.8.0,<3.0.0",
    "litellm>=1.77.3",
    "asyncio>=3.4.3",
    "structlog>=24.1.0",
    "pypdf>=5.3.0",
    "python-docx>=1.1.2",
    "tenacity>=8.2.3",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
pdf = ["pdf2image>=1.17.0"]
docx = ["python-docx>=1.1.0"]
all = ["pdf2image>=1.17.0", "python-docx>=1.1.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
    "black>=24.2.0",
    "mypy>=1.8.0",
    "ruff>=0.2.1",
    "isort>=5.13.2",
    "pdf2image>=1.17.0",
    "python-docx>=1.1.0",
]

[project.scripts]
tinbox = "tinbox.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/tinbox"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --cov=tinbox"
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_optional = true

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
